#10074: Move reciprocal_bw to TTNN
