+++ tt_eager/tt_dnn/op_library/backward/backward_ops.cpp
@@ -479,29 +479,6 @@ std::vector<Tensor> deg2rad_bw(const Tensor& grad, const Tensor& input, const Me
     return operation::decorate_as_composite(__func__, _deg2rad_bw)(grad, input, output_mem_config);
 }
 
-
-std::vector<Tensor> _reciprocal_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    Tensor t_inf = full_like(input, std::numeric_limits<float>::infinity(), output_mem_config);
-    Tensor t_nan = full_like(input, std::nanf(""), output_mem_config);
-    grad_tensor.emplace_back(where(
-        ttnn::eqz(input, output_mem_config),
-        where(
-            ttnn::eqz(grad, output_mem_config),
-            t_nan,
-            ttnn::multiply(t_inf, ttnn::neg(ttnn::sign(grad, output_mem_config), output_mem_config), std::nullopt, output_mem_config),
-            output_mem_config),
-        ttnn::multiply(ttnn::neg(grad, output_mem_config),
-            ttnn::reciprocal(ttnn::square(input, output_mem_config), output_mem_config),
-            std::nullopt,
-            output_mem_config),
-        output_mem_config));
-    return grad_tensor;
-}
-std::vector<Tensor> reciprocal_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _reciprocal_bw)(grad, input, output_mem_config);
-}
-
 // Autoformat support
 Tensor change_layout_to_tile(const Tensor& temp, const MemoryConfig& output_mem_config) {
     auto formatted_input_tensor = temp;
