+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -167,11 +167,6 @@ std::vector<Tensor> deg2rad_bw(
     const Tensor& input,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
-std::vector<Tensor> reciprocal_bw(
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
-
 std::vector<Tensor> prod_bw(
     const Tensor& grad,
     const Tensor& input,
